+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Include your changelog here for each version being deployed
 
+## v0.0.4
+- Most bugfixes in some classes
+- Using new explainer for deep models
+
 ## v0.0.1
 
 - Adding regression and classification trees based on xGBoost